fix(ui): keep the BIU outside-temp offset decode signed (step 6g-2)
The step 6g-2 clang-tidy signed-bitwise cleanup added an unsigned suffix
to every mask in this decode, which turned the whole expression into
unsigned arithmetic: the final "- 4" wrapped around instead of going
negative for nibbles 0xC-0xF, so negative outside-temperature offsets
displayed as ~2.1e9 instead of small negative floats. Do the masking in
unsigned, but cast back to int before the biasing subtraction so the
result can go negative again, matching pre-cleanup behavior.

+++ src/ui/desktop/biu/biu_operations_subaru.cpp
@@ -999,7 +999,7 @@ void BiuOperationsSubaru::parse_biu_message(const QByteArray& message)
             {
                 temp = biu_tt_names.at(4);
                 biu_tt_result->append((uint8_t)message.at(7) & 0x0FU);
-                calc_result = (((((uint8_t)message.at(7) & 0x0FU) + 4) & 0x0FU) - 4) * 0.5;
+                calc_result = (static_cast<int>((((uint8_t)message.at(7) & 0x0FU) + 4U) & 0x0FU) - 4) * 0.5;
                 temp.append(QString("%1 ").arg(calc_result));
                 temp.append(biu_tt_names.at(5));
                 data_result->append(temp);
